refactor : FE websocket response 맞추기
