+++ src/main/java/com/mopl/api/domain/user/dto/response/WatchingSessionDto.java
@@ -1,7 +1,6 @@
 package com.mopl.api.domain.user.dto.response;
 
 import com.mopl.api.domain.content.dto.response.ContentDto;
-import com.mopl.api.domain.user.entity.WatchingSession;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.Builder;
@@ -10,15 +9,8 @@ import lombok.Builder;
 public record WatchingSessionDto(
     UUID id,
     LocalDateTime createdAt,
-    UserDto user,
+    UserSummary watcher,
     ContentDto content
 ) {
-    public static WatchingSessionDto from(WatchingSession ws) {
-        return WatchingSessionDto.builder()
-                                 .id(ws.getId())
-                                 .createdAt(ws.getCreatedAt())
-                                 .user(UserDto.from(ws.getWatcher()))
-                                 .content(ContentDto.from(ws.getContent()))
-                                 .build();
-    }
+
 }
